pyproject.toml
[project]
name = "stable-diff-fryan"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.12.0",
    "bitsandbytes>=0.48.2",
    "cog>=0.16.8",
    "diffusers>=0.36.0",
    "protobuf>=6.33.2",
    "python-dotenv>=1.0.0",
    "sentencepiece>=0.2.1",
    "torch==2.8.0",
    "transformers>=4.57.3",
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[dependency-groups]
dev = [
    "ipywidgets>=8.1.8",
    "jupyterlab>=4.5.1",
]
